Centralize unproblematic EventData structs in events package.
