+++ lib/model/folder.go
@@ -341,19 +341,6 @@ func (f *folder) getHealthErrorWithoutIgnores() error {
 	return nil
 }
 
-type ItemStartedEventData struct {
-	Folder string `json:"folder"`
-	Item   string `json:"item"`
-	Type   string `json:"type"`
-	Action string `json:"action"`
-}
-
-type ItemFinishedEventData struct {
-	ItemStartedEventData
-
-	Error *string `json:"error,omitempty"`
-}
-
 func (f *folder) pull() (success bool, err error) {
 	f.pullFailTimer.Stop()
 	select {
@@ -1070,12 +1057,6 @@ func (f *folder) monitorWatch(ctx context.Context) {
 	}
 }
 
-type FolderWatchStateChangedEventData struct {
-	Folder string  `json:"folder"`
-	From   *string `json:"from,omitempty"`
-	To     *string `json:"to,omitempty"`
-}
-
 // setWatchError sets the current error state of the watch and should be called
 // regardless of whether err is nil or not.
 func (f *folder) setWatchError(err error, nextTryIn time.Duration) {
@@ -1084,7 +1065,7 @@ func (f *folder) setWatchError(err error, nextTryIn time.Duration) {
 	f.watchErr = err
 	f.watchMut.Unlock()
 	if err != prevErr {
-		data := FolderWatchStateChangedEventData{
+		data := events.FolderWatchStateChangedEventData{
 			Folder: f.ID,
 		}
 		if prevErr != nil {
@@ -1228,14 +1209,6 @@ func (f *folder) updateLocalsFromPulling(fs []protocol.FileInfo) {
 	f.emitDiskChangeEvents(fs, events.RemoteChangeDetected)
 }
 
-type LocalIndexUpdatedEventData struct {
-	Folder    string   `json:"folder"`
-	Items     int      `json:"items"`
-	Filenames []string `json:"filenames"`
-	Sequence  int64    `json:"sequence"`
-	Version   int64    `json:"version"` // DEPRECATED: legacy for Sequence
-}
-
 func (f *folder) updateLocals(fs []protocol.FileInfo) {
 	f.fset.Update(protocol.LocalDeviceID, fs)
 
@@ -1249,7 +1222,7 @@ func (f *folder) updateLocals(fs []protocol.FileInfo) {
 	f.forcedRescanPathsMut.Unlock()
 
 	seq := f.fset.Sequence(protocol.LocalDeviceID)
-	f.evLogger.Log(events.LocalIndexUpdated, LocalIndexUpdatedEventData{
+	f.evLogger.Log(events.LocalIndexUpdated, events.LocalIndexUpdatedEventData{
 		Folder:    f.ID,
 		Items:     len(fs),
 		Filenames: filenames,
@@ -1257,18 +1230,6 @@ func (f *folder) updateLocals(fs []protocol.FileInfo) {
 		Version:   seq, // DEPRECATED: legacy for Sequence
 	})
 }
-
-type DiskChangeDetectedEventData struct {
-	Folder     string `json:"folder"`
-	FolderID   string `json:"folderID"` // incorrect, deprecated, kept for historical compliance
-	Label      string `json:"label"`
-	Action     string `json:"action"`
-	Type       string `json:"type"`
-	Path       string `json:"path"`
-	ModifiedBy string `json:"modifiedBy"`
-}
-type LocalChangeDetectedEventData DiskChangeDetectedEventData
-type RemoteChangeDetectedEventData DiskChangeDetectedEventData
 
 func (f *folder) emitDiskChangeEvents(fs []protocol.FileInfo, typeOfEvent events.EventType) {
 	for _, file := range fs {
@@ -1290,7 +1251,7 @@ func (f *folder) emitDiskChangeEvents(fs []protocol.FileInfo, typeOfEvent events
 		}
 
 		// Two different events can be fired here based on what EventType is passed into function
-		f.evLogger.Log(typeOfEvent, DiskChangeDetectedEventData{
+		f.evLogger.Log(typeOfEvent, events.DiskChangeDetectedEventData{
 			Folder:     f.ID,
 			FolderID:   f.ID, // incorrect, deprecated, kept for historical compliance
 			Label:      f.Label,
